move question 10 to topof part 2 and make it question 12
